python/pyproject.toml
[project]
name = "superlamp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "langchain>=1.2.7",
    "langchain-chroma>=1.1.0",
    "langchain-openai>=1.1.7",
    "langfuse>=3.12.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "beautifulsoup4>=4.14.3",
    "langchain-community>=0.4.1",
    "langchain-tavily>=0.2.17",
    "lxml>=6.0.2",
]
